NEW: GameObjects can now be rotated and scaled (using the Transform)
- Bug fix: Modified the Transform.rotation field to double
- Added a Vector2 Transform.scale field
- Modified Texture and RenderManager to use the transform and therein the scale and rotations for rendering

+++ SDL Game/SDL Game/Engine/Vector2.cpp
@@ -2,13 +2,13 @@
 #include <cmath>
 
 
-float Vector2::getLength()
+double Vector2::getLength()
 {
-	return sqrtf(getLengthSquared());
+	return sqrt(getLengthSquared());
 }
 
 
-float Vector2::getLengthSquared()
+double Vector2::getLengthSquared()
 {
 	return x * x + y * y;
 }
